LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/open-url-redirected-worker-origin-expected.txt

Blocked access to external URL http://www2.localhost:8800/XMLHttpRequest/resources/inspect-headers.py?filter_name=origin&cors

PASS XMLHttpRequest: redirected worker scripts, origin and referrer 
FAIL Origin header assert_equals: expected "Origin: http://localhost:8800\n" but got ""
FAIL Referer header assert_equals: expected "Referer: http://localhost:8800/XMLHttpRequest/resources/workerxhr-origin-referrer.js\n" but got "Referer: http://localhost:8800/XMLHttpRequest/resources/redirect.py?location=workerxhr-origin-referrer.js\n"
PASS Request URL test 
